+++ pybop/parameters/parameter.py
@@ -11,6 +11,9 @@ import scipy.stats as stats
 from numpy.typing import NDArray
 
 from pybop.parameters.distributions import Distribution
+from pybop.parameters.multivariate_distributions import (
+    MarginalDistribution,
+)
 from pybop.transformation.base_transformation import Transformation
 from pybop.transformation.transformations import (
     ComposedTransformation,
@@ -117,6 +120,10 @@ class Parameter:
         self._bounds = None
         self._transformation = transformation or IdentityTransformation()
 
+        # The ep-bolfi optimiser requires some distribution properties in the search space rather than the model space
+        # Some transformations are not suitable for some multivariate distributions as they are currently implemented in this context
+        self._check_compatible_transformation()
+
         if self._distribution is not None:
             lower, upper = self._distribution.support()
             if np.isinf(lower) and np.isinf(upper):
@@ -147,6 +154,20 @@ class Parameter:
 
         # Validate initial values are within bounds
         self._validate_values_within_bounds()
+
+    def _check_compatible_transformation(self):
+        if isinstance(self._distribution, MarginalDistribution):
+            allowed_transformations = (
+                self._distribution.parent_distribution.compatible_transformations
+            )
+
+            if not isinstance(self._transformation, allowed_transformations):
+                raise TypeError(
+                    f"The transformation provided is not compatible with pybop.{self._distribution.parent_distribution.name}. "
+                    "Only "
+                    + ", ".join([trans.__name__ for trans in allowed_transformations])
+                    + " are allowed."
+                )
 
     def sample_from_distribution(
         self,
@@ -253,12 +274,12 @@ class Parameters:
             raise TypeError(
                 "parameters must be either a dictionary or a pybop.Parameters instance"
             )
-
         self._parameters = OrderedDict()
         for name, param in parameters.items():
-            self._add(name, param, update_transform=False)
+            self._add(name, param, update_transform=False, check_multivariate=False)
 
         self._transform = self.construct_transformation()
+        self.check_multivariate()
 
     def __getitem__(self, name: str) -> Parameter:
         return self.get(name)
@@ -280,12 +301,46 @@ class Parameters:
     def __iter__(self) -> Iterator[Parameter]:
         return iter(self._parameters.values())
 
-    def add(self, name: str, parameter: Parameter) -> None:
+    def add(self, name: str, parameter: Parameter, check_multivariate=True) -> None:
         """Add a parameter to the collection."""
-        self._add(name, parameter)
+        self._add(name, parameter, check_multivariate=check_multivariate)
+
+    def check_multivariate(self):
+        """Method to determin whether parameters have a MultivariateDistribution
+        Multivariate distributions are passed to individual parameters via the corresponding marginal distribution.
+        The pybop.MarginalDistribution class retains the underlying pybop.MultivariateDistribution in the parent_distribution property
+        """
+
+        # check if any distribution is a pybop.MarginalDistribution
+        self._multivariate = any(
+            isinstance(param.distribution, MarginalDistribution) for param in self
+        )
+
+        # if there is a pybop.MarginalDistribution ensure all distributions are marginal distributions of the same parent_distribution
+        if self._multivariate:
+            if not all(
+                isinstance(param.distribution, MarginalDistribution) for param in self
+            ):
+                raise TypeError(
+                    "A Parameters object with a MarginalDistribution cannot be combined with parameters with other types of distributions"
+                )
+            dist = next(
+                iter(self._parameters.values())
+            ).distribution.parent_distribution
+            if not all(
+                param.distribution.parent_distribution == dist for param in self
+            ):
+                raise ValueError(
+                    "All MarginalDistributions must share the same parent MultivariateDistribution."
+                )
+            self.distribution = dist
 
     def _add(
-        self, name: str, parameter: Parameter, update_transform: bool = True
+        self,
+        name: str,
+        parameter: Parameter,
+        update_transform: bool = True,
+        check_multivariate=True,
     ) -> None:
         """
         Internal method to add a parameter to the collection.
@@ -307,6 +362,8 @@ class Parameters:
 
         if update_transform:
             self._transform = self.construct_transformation()
+        if check_multivariate:
+            self.check_multivariate()
 
     def remove(self, name: str) -> Parameter:
         """Remove parameter and return it."""
@@ -326,9 +383,13 @@ class Parameters:
         """
         for name, param in parameters.items():
             if name not in self._parameters.keys():
-                self.add(name, param)
+                self.add(
+                    name, param, check_multivariate=False
+                )  # don't check every each param individually
             else:
                 print(f"Discarding duplicate {name}.")
+
+        self.check_multivariate()  # check once when all parameters are added
 
     def get(self, name: str) -> Parameter:
         """Get a parameter by name."""
@@ -336,13 +397,15 @@ class Parameters:
             raise ParameterNotFoundError(f"Parameter for '{name}' not found")
         return self._parameters[name]
 
-    def set(self, name: str, param: Parameter) -> None:
+    def set(self, name: str, param: Parameter, check_multivariate=True) -> None:
         """Get a parameter by name."""
         if name not in self._parameters:
             raise ParameterNotFoundError(f"Parameter for '{name}' not found")
         if not isinstance(param, Parameter):
-            raise TypeError({"Paremeter must be of type pybop.ParemterInfo"})
+            raise TypeError({"Paremeter must be of type pybop.Parameter"})
         self._parameters[name] = param
+        if check_multivariate:
+            self.check_multivariate()
 
     def get_bounds(self, transformed: bool = False) -> dict:
         """
@@ -449,22 +512,51 @@ class Parameters:
         """
         Sample from each parameter distribution.
 
+        or
+
+        Draw random samples from the joint parameters distribution for multivariate parameters.
+
+        Parameters
+        ----------
+        n_samples : int
+            The number of samples to draw (default: 1).
+        random_state : int, optional
+            The random state seed for reproducibility (default: None).
+        transformed: bool
+            If True, the transformation is applied to the output
+            (default: False).
+
         Returns
         -------
         NDArray[np.floating] or None
             Array of shape (n_samples, n_parameters) or None if any distribution is missing
         """
-        all_samples = []
 
-        for param in self._parameters.values():
-            samples = param.sample_from_distribution(
-                n_samples, random_state=random_state, transformed=transformed
-            )
-            if samples is None:
-                return None
-            all_samples.append(samples)
+        if self._multivariate:
+            # use multivariate distribution for to sample all parameters
+            samples = self.distribution.rvs(n_samples, random_state=random_state)
+            if samples.ndim < 2:
+                samples = np.atleast_2d(samples)
 
-        return np.column_stack(all_samples)
+            if transformed:
+                samples = np.asarray(
+                    [self.transformation.to_search(s) for s in samples]
+                )
+
+            return samples
+        else:
+            # sample each parameter individually
+            all_samples = []
+
+            for param in self._parameters.values():
+                samples = param.sample_from_distribution(
+                    n_samples, random_state=random_state, transformed=transformed
+                )
+                if samples is None:
+                    return None
+                all_samples.append(samples)
+
+            return np.column_stack(all_samples)
 
     def get_sigma0(self, transformed: bool = False) -> list:
         """
@@ -557,6 +649,15 @@ class Parameters:
             return None
 
         return ComposedTransformation(transformations)
+
+    @property
+    def transformed_distribution_properties(self):
+        # retrieve properties of the distribution in the search space
+        # needed for ep-bolfi optimiser
+        if self._multivariate:
+            return self.distribution.transformed_properties(self._transform)
+        else:
+            raise NotImplementedError
 
     def get_bounds_for_plotly(self, transformed: bool = False) -> np.ndarray:
         """